Split up the rest of the old blake2b module
