+++ tests/hash/mod.rs
@@ -5,7 +5,7 @@ pub mod sha384_nist_cavp;
 pub mod sha512_nist_cavp;
 
 use crate::TestCaseReader;
-use orion::hazardous::hash::{self, sha2::sha256, sha2::sha384, sha2::sha512};
+use orion::hazardous::hash::{blake2, sha2::sha256, sha2::sha384, sha2::sha512};
 use orion::hazardous::mac;
 
 fn blake2b_test_runner(input: &[u8], key: &[u8], output: &[u8]) {
